swithced to  np.longdouble

+++ src/libSOGAtruncate.py
@@ -14,10 +14,10 @@ import timing
 
 # Check if np.float128 is available
 if hasattr(np, 'longdouble'):
-    sogaType=np.longdouble
+    sogaType=np.float64
 else:
     print("np.float128 not available swithing to np.float64")
-    sogaType=np.longdouble
+    sogaType=np.float64
 
 def negate(trunc):
     """ Produces a string which is the logic negation of trunc """
